パイプに対応した
