+++ execution.h
@@ -24,6 +24,10 @@ char	*find_executable_file_from_path_env(char *filename);
 char	*find_executable_file_in_cwd(char *filename);
 int		ft_execvp(char *filename, char **argv);
 int		command_execution(t_command_invocation *command);
+int		set_input_file(t_command_invocation *command);
+int		set_output_file(t_command_invocation *command);
+int		pipe_process(t_command_invocation *command, int pipe_fd[2]);
+int		spawn_child(t_command_invocation *command);
 char	*path_join(char *dirpath, char *filename);
 
 #endif
